backend/requirements.txt
# FastAPI and server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-dotenv==1.0.1
pydantic==2.10.4

# AI and LLM
google-generativeai==0.8.3
langgraph==0.2.63
langchain==0.3.18
langchain-google-genai==2.0.7

# Utilities
aiohttp==3.11.11
httpx==0.28.1
